refactor: full Web/WASM compatibility (conditional imports, web stubs, no dart:io)
