+++ lib/src/core/storage/native_storage_adapter_web.dart
@@ -1,0 +1,6 @@
+/// Web/WASM stub for NativeStorageAdapter
+class NativeStorageAdapter {
+  static Future<String> getDocumentsDirectory() async => '';
+  static Future<String> getSupportDirectory() async => '';
+  static Future<String> getCacheDirectory() async => '';
+}
